Добавлен Visualizer.
